+++ FarmFarm/src/main/java/edu/kh/farmfarm/board/controller/BoardWriteController.java
@@ -1,10 +1,12 @@
 package edu.kh.farmfarm.board.controller;
 
 import java.util.List;
+import java.util.Map;
 
 import org.apache.logging.log4j.core.config.plugins.validation.constraints.Required;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
+import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestParam;
@@ -20,17 +22,18 @@ public class BoardWriteController {
 	@Autowired
 	private BoardWriteService serivce;
 	
-	@GetMapping("/board/write")
-	private String boardWritePage() {
-		return "board/boardWrite";
-	}
-	
-	@PostMapping("/board/write")
-	private String boardWrite(
-			Board board,
-			@RequestParam(value="imgs", required = false) List<MultipartFile> imgList) {
-		return "redirect:board";
-	}
+//	@GetMapping("/board/write")
+//	private String boardWritePage(Model model) {
+//		return "board/boardWrite";
+//	}
+//	
+//	@PostMapping("/board/write")
+//	private String boardWrite(
+//			Board board,
+//			@RequestParam(value="imgs", required = false) List<MultipartFile> imgList) {
+//		
+//		return "redirect:board";
+//	}
 	
 
 }
