+++ FarmFarm/src/main/java/edu/kh/farmfarm/board/model/dao/BoardDAO.java
@@ -1,0 +1,20 @@
+package edu.kh.farmfarm.board.model.dao;
+
+import java.util.List;
+import java.util.Map;
+
+import org.mybatis.spring.SqlSessionTemplate;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public class BoardDAO {
+
+	@Autowired
+	private SqlSessionTemplate sqlSession;
+	
+	public List<Map<String, Object>> boardTypeList() {
+		return sqlSession.selectList("boardMapper.boardTypeList");
+	}
+
+}
