+++ FarmFarm/src/main/java/edu/kh/farmfarm/board/model/service/BoardService.java
@@ -1,0 +1,10 @@
+package edu.kh.farmfarm.board.model.service;
+
+import java.util.List;
+import java.util.Map;
+
+public interface BoardService {
+
+	List<Map<String, Object>> boardTypeList();
+
+}
