+++ FarmFarm/src/main/java/edu/kh/farmfarm/board/model/service/BoardServiceImpl.java
@@ -1,0 +1,22 @@
+package edu.kh.farmfarm.board.model.service;
+
+import java.util.List;
+import java.util.Map;
+
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.stereotype.Service;
+
+import edu.kh.farmfarm.board.model.dao.BoardDAO;
+
+@Service
+public class BoardServiceImpl implements BoardService {
+	
+	@Autowired
+	private BoardDAO dao;
+
+	@Override
+	public List<Map<String, Object>> boardTypeList() {
+		return  dao.boardTypeList();
+	}
+
+}
