+++ FarmFarm/src/main/webapp/WEB-INF/views/board/boardList.jsp
@@ -27,6 +27,11 @@
                 <span>팁</span>
                 <span>질문</span>
             </div>
+            <%-- <div class="board-nav-area">
+                <c:forEach var="boardType" items="${boardTypeList}">
+                    <a href="/board/${boardType.BOARD_TYPE_NO}">${boardType.BOARD_NAME}</a>
+                </c:forEach>
+            </div> --%>
         </section>
         <section class="board-search">
             <form action="" class="board-search-area">
@@ -74,6 +79,48 @@
                 </ul>
                 <div class="board-write-bottom">
                     <div class="board-pagination">
+                                <div class="pagination-area">
+                <%-- <!-- ajax로 만들어 보십시다 -->
+                <div class="page-box">
+                  <i class="fa-solid fa-angles-left"></i>
+                </div>
+                <div class="page-box">
+                  <i class="fa-solid fa-angle-left"></i>
+                </div>
+                <div class="page-box">
+                    1
+                </div>
+                <div class="page-box">
+                    2
+                </div>
+                <div class="page-box">
+                    3
+                </div>
+                <div class="page-box">
+                    4
+                </div>
+                <div class="page-box">
+                    5
+                </div>
+                <div class="page-box">
+                    6
+                </div>
+                <div class="page-box">
+                    7
+                </div>
+                <div class="page-box">
+                    8
+                </div>
+                <div class="page-box">
+                    9
+                </div>
+                <div class="page-box">
+                  <i class="fa-solid fa-angle-right"></i>
+                </div>
+                <div class="page-box">
+                  <i class="fa-solid fa-angles-right"></i>
+                </div>
+            </div> --%>
                         <ul class="pagination">
 
                             <%-- 첫 페이지 이동 --%>
