+++ FarmFarm/src/main/webapp/WEB-INF/views/board/boardWrite.jsp
@@ -26,19 +26,19 @@
                 <div class="board-category-area">
                     <span class="category">카테고리 선택</span>
 
-                    <%-- <input type="radio" name="boardTypeNo" id="share" value="1">
+                    <input type="radio" name="boardTypeNo" id="share" value="1">
                     <label for="share">물물교환</label>
 
                     <input type="radio" name="boardTypeNo" id="tip" value="2">
                     <label for="tip">팁</label>
 
                     <input type="radio" name="boardTypeNo" id="question" value="3">
-                    <label for="question">질문</label> --%>
+                    <label for="question">질문</label>
 
-                    <c:forEach var="boardType" items="${boardTypeList}">
-                        <input type="radio" id="${boardType.boardNo}">
-                        <label for="${boardType.boardNo}">${boardType.boardName}</label>
-                    </c:forEach>
+                    <%-- <c:forEach var="boardType" items="${boardTypeList}">
+                        <input type="radio" name="boardTypeNo" id="${boardType.BOARD_TYPE_NO}">
+                        <label for="${boardType.BOARD_TYPE_NO}">${boardType.BOARD_NAME}</label>
+                    </c:forEach> --%>
 
                 </div>
                 <button class="board-submit">게시하기</button>
